+++ great_expectations/exceptions.py
@@ -65,10 +65,6 @@ class UnsupportedConfigVersionError(GreatExpectationsError):
     pass
 
 
-class ZeroDotSevenConfigVersionError(GreatExpectationsError):
-    pass
-
-
 class ProfilerError(GreatExpectationsError):
     pass
 
@@ -76,6 +72,14 @@ class ProfilerError(GreatExpectationsError):
 class InvalidConfigError(DataContextError):
     def __init__(self, message):
         self.message = message
+
+
+class MissingConfigVariableError(InvalidConfigError):
+    def __init__(self, message, missing_config_variable=None):
+        if not missing_config_variable:
+            missing_config_variable = []
+        self.message = message
+        self.missing_config_variable = missing_config_variable
 
 
 class AmbiguousDataAssetNameError(DataContextError):